Final Version Pre-Deployment

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteDataServlet.java
@@ -14,15 +14,23 @@
  
 package com.google.sps.servlets;
  
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
-import com.google.appengine.api.datastore.Key;
-import com.google.appengine.api.datastore.KeyFactory;
 import java.io.IOException;
+import com.google.sps.data.Book;
+import com.google.gson.Gson;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import com.google.appengine.api.datastore.DatastoreService;
+import com.google.appengine.api.datastore.DatastoreServiceFactory;
+import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.PreparedQuery;
+import com.google.appengine.api.datastore.Query;
+import com.google.appengine.api.datastore.Query.SortDirection;
+import com.google.appengine.api.datastore.Key;
+import com.google.appengine.api.datastore.KeyFactory;
+import com.google.gson.Gson;
+
 
 /** Servlet that deletes book data no longer in use*/
 @WebServlet("/delete-data")
@@ -30,11 +38,19 @@ public final class DeleteDataServlet extends HttpServlet {
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-      
-    int id = Integer.parseInt(request.getParameter("id"));
-    Key bookEntityKey = KeyFactory.createKey("Book", id);
+
+    Query query = new Query("Book");
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-    datastore.delete(bookEntityKey);
+    PreparedQuery results = datastore.prepare(query);
+
+    for(Entity entity : results.asIterable()){
+        String book_title = (String) entity.getProperty("title");
+        String author = (String) entity.getProperty("author");
+        if(book_title == null || author == null){
+            Key bookEntityKey = KeyFactory.createKey("Book", book_title);
+            datastore.delete(bookEntityKey);
+        }
+    }
   }
 }
